pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "support-ops-openenv"
version = "0.1.0"
description = "A real-world OpenEnv customer support operations environment."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "openai>=1.68.0",
  "openenv-core>=0.1.0",
  "pydantic>=2.10.0",
  "pyyaml>=6.0.2",
  "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0"]

[tool.setuptools.packages.find]
include = ["support_ops_env*"]
